refactor: 패키지 이동 및 companyDeliveryManagerId -> companyDeliveryPersonId 필드명 변경
